setup.py now installs the 'tests' directory with the package.
When I tried to install a package that depends on Concrete, I was
getting the error message:

  AttributeError: 'module' object has no attribute 'test_conversion'

The 'tests/test_conversion.py' script exists only in the Concrete
repo, so it appears that the other package's setup.py script tried to
run the unit tests for it's dependencies.

Added the 'tests' directory to the list of directories that are
installed with the concrete package.  This fixed the dependency issue
with the other module.