Error handling for concrete_type is now handled by argparse.
The concrete_type command line option specifies the type of Concrete
object to convert to JSON.  There is a whitelist of possible types,
and the two current options are 'communication' and 'tokenlattice'.
The script was using some custom code to check if the value for
'concrete_type' was on the whitelist, but this checking is now handled
by argparse.