Resolve undefined variables
Python, unlike Java, only defines and uses instance variables when
they are prefixed by 'self' (or the reference to the instance). In
several places, an instance variable was defined but a local variable
was being used or vice versa.

This change cleans up the variable definitions so they are all defined
before their use.