Allow reuse of AnnotatorClientWrapper
The current implementation of the AnnotatorClientWrapper opens the
transport when the object is instantiated (i.e., in the __init__
method), which is undesirable for two reasons. First, there is no
guarantee that the transport is closed promptly (or at all). Second,
the same AnnotatorClientWrapper cannot be used in multiple with blocks
because the first exit closes the transport.

It is preferable to open the transport as part of the __enter__ method
to address both issues. If __enter__ opens the transport, then
__exit__ will close it automatically, which avoids any resource leaks.
In addition, the same object can be reused multiple times without
maintaining an active network connection -- e.g.,
    wrapper = AnnotatorClientWrapper(host, port)
    with wrapper:
        wrapper.annotate(comm)
    ...
    with wrapper:
        wrapper.annotate(comm)
This change does not make the AnnotatorClientWrapper reentrant, as
nesting it is not permitted.

See Python's contextlib library for more information about single use,
reusable, and reentrant context managers: https://goo.gl/ITZCqK