Support all Annotator operations
This change modifies AnnotatorClientWrapper's __enter__ method to
return an Annotator.Client. It allows the other operations (i.e.,
getMetadata, getDocumentation, and shutdown) defined by the Annotator
interface to be invoked instead of just the annotate method. The
annotate method is also removed because annotate can be invoked
directly on the Annotator.Client instance.

An alternative to this change would be adding the aforementioned
methods to the AnnotatorClientWrapper. Such a change would make the
wrapper a proxy for the Annotator.Client instance, which seems like
unnecessary indirection without additional functionality.