Fix TypeError when creating error message
Tokenization.kind is assumed to be an enumeration that is represented
as an integer, but Python does not provide strong type guarantees. If
a string or another type is provided instead, then creation of the
error message will itself raise a TypeError.

This change resolves this issue by using '%s' as the conversion string
rather than '%d'. A unit test is included to prevent regressions.