Extract Thrift dependencies to single module
Thrift dependencies in the Concrete code are common, but such
dependencies make maintenance more difficult because many different
modules must be changed when modifying an implementation decision
(e.g., which protocol to use as the default).

This change creates a thrift_factory module to encapsulate all such
concretizations of Thrift objects.