Enforce Flake8 stylistic checks
Flake8 is a wrapper for several different tools, including PyFlakes,
pep8, and a McCabe complexity checker. Many of the checks are
stylistic although common coding mistakes (e.g., using an undefined
variable) are also reported.

This change enables Flake8 by default as part of the tox tests but
only for the concrete/util module. Other modules are omitted because
they are auto-generated by the Thrift compiler and do not adhere to
Python's standard coding style. (It would be advantageous to restrict
all Thrift code to a single module (such as concrete.thrift) to allow
all other code to be checked more easily.)